requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
decorator==4.4.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
logger==1.4
MarkupSafe==1.1.1
pycparser==2.20
pyTelegramBotAPI==3.7.3
python-decouple==3.4
python-telegram-bot==13.1
pytz==2020.5
requests==2.25.1
schedule==0.6.0
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
